Enum cvar::InnerError

pub enum InnerError {
    NameError,
    ListError,
    PropError,
    ActionError,
    ParseError(Box<StdError>),
    ChangeError(Box<StdError>),
    ConstError,
    InvokeError(Box<StdError>),
}

Contextless error.

Variants
NameError

Name not found error.

When traversing the cvar hierarhcy, a child by the name was not found.

ListError

Node is not a list.

When traversing the cvar hierarchy, expected the child to implement IList.

This happens when an id is given (eg. foo.prop.baz) but foo.prop is not a list of cvars itself.

PropError

Node is not a property.

When traversing the cvar hierarchy, expected the child to implement IProperty.

This happens when an id is given (eg. foo.list) to get or set its value but foo.list is not a property.

ActionError

Node is not an action.

When traversing the cvar hierarchy, expected the child to implement IAction.

This happens when an id is invoked (eg. foo.bar) but foo.bar is not an action.

ParseError(Box<StdError>)

Error parsing the value.

ChangeError(Box<StdError>)

Error validating the value.

ConstError

Cannot modify the cvar.

The property is read-only and cannot be modified.

InvokeError(Box<StdError>)

Error invoking the action.