Struct cvar::Action

pub struct Action<N, D, F> where N: Borrow<str>, D: Borrow<str>, F: OnInvoke {
    // some fields omitted
}

Action instance.

The Name and Description types allow abstracting over &'static str, &'a str and String. This supports dynamic cvars while only paying for what you need.

F is the callable type called when the action is invoked.

Methods

impl<N, D> Action<N, D, Pass> where N: Borrow<str>, D: Borrow<str>

fn new<F>(name: N, desc: D, f: F) -> Action<N, D, F> where F: Fn(&mut Context) -> BoxResult<()>

Creates a new Action.

Given a name, description and a callback to be invoked.