Struct cvar::RefProperty

pub struct RefProperty<'a, T, F> where T: 'a + Value, F: OnChange<T> {
    // some fields omitted
}

Property instance of Clone data.

Methods

impl<'a, T> RefProperty<'a, T, ()> where T: 'a + Value

fn new<I: Into<T>>(name: &'a str, desc: &'a str, val: &'a RefCell<T>, def: I) -> RefProperty<'a, T, ()>

Creates a new RefProperty.

Given a name, description, a reference to the underlying value and an appropriate default.

Example

// The underlying data wrapped in a `RefCell`.
use std::cell::RefCell;
let var = RefCell::new(String::from("Hello, World!"));

// The variable wrapped in a `RefProperty`.
use cvar::{RefProperty, IProperty};
let prop = RefProperty::new("prop", "property description", &var, "Hello, Default!");
assert_eq!(prop.get().unwrap(), "Hello, World!");
prop.reset();
assert_eq!(*var.borrow(), "Hello, Default!");

impl<'a, T, F> RefProperty<'a, T, F> where T: 'a + Value, F: OnChange<T>

fn checked<I: Into<T>>(name: &'a str, desc: &'a str, val: &'a RefCell<T>, def: I, change: F) -> RefProperty<'a, T, F>

Creates a new Property with an onchange callback.

Given a name, description, a reference to the underlying value, an appropriate default and a change callback.

The change callback is called with a reference to the old value and a the new value after it has successfully parsed.